+++ src/Interfaces/ManyRelationInterface.php
@@ -3,6 +3,7 @@
 namespace Swis\JsonApi\Client\Interfaces;
 
 use Swis\JsonApi\Client\Collection;
+use Swis\JsonApi\Client\Meta;
 
 interface ManyRelationInterface
 {
@@ -39,4 +40,14 @@ interface ManyRelationInterface
      * @return bool
      */
     public function shouldOmitIncluded(): bool;
+
+    /**
+     * @param \Swis\JsonApi\Client\Meta|null $meta
+     */
+    public function setMeta(Meta $meta = null);
+
+    /**
+     * @return \Swis\JsonApi\Client\Meta|null
+     */
+    public function getMeta();
 }