+++ src/Interfaces/OneRelationInterface.php
@@ -2,6 +2,8 @@
 
 namespace Swis\JsonApi\Client\Interfaces;
 
+use Swis\JsonApi\Client\Meta;
+
 interface OneRelationInterface
 {
     /**
@@ -37,4 +39,14 @@ interface OneRelationInterface
      * @return bool
      */
     public function shouldOmitIncluded(): bool;
+
+    /**
+     * @param \Swis\JsonApi\Client\Meta|null $meta
+     */
+    public function setMeta(Meta $meta = null);
+
+    /**
+     * @return \Swis\JsonApi\Client\Meta|null
+     */
+    public function getMeta();
 }