+++ src/Item.php
@@ -443,10 +443,11 @@ class Item extends Model implements ItemInterface
      *
      * @param string                                        $relation
      * @param \Swis\JsonApi\Client\Interfaces\DataInterface $value
+     * @param \Swis\JsonApi\Client\Meta|null                $meta
      *
      * @return static
      */
-    public function setRelation(string $relation, DataInterface $value)
+    public function setRelation(string $relation, DataInterface $value, Meta $meta = null)
     {
         if (method_exists($this, $relation)) {
             /** @var \Swis\JsonApi\Client\Interfaces\OneRelationInterface|\Swis\JsonApi\Client\Interfaces\ManyRelationInterface $relationObject */
@@ -458,6 +459,7 @@ class Item extends Model implements ItemInterface
         }
 
         $relationObject->associate($value);
+        $relationObject->setMeta($meta);
 
         return $this;
     }