+++ src/Relations/AbstractRelation.php
@@ -2,8 +2,12 @@
 
 namespace Swis\JsonApi\Client\Relations;
 
+use Swis\JsonApi\Client\Traits\HasMeta;
+
 abstract class AbstractRelation
 {
+    use HasMeta;
+
     /**
      * @var \Swis\JsonApi\Client\Interfaces\DataInterface|null
      */