+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Added meta to `ItemInterface`.
 N.B. This is a breaking change if you implement the `ItemInterface` yourself instead of using the supplied `Item`.
 * Added `Jsonapi` class.
+* Added meta to `OneRelationInterface` and `ManyRelationInterface`.
+N.B. This is a breaking change if you implement (one of) these interfaces yourself instead of using the supplied relations.
 
 ### Changed
 
@@ -20,6 +22,7 @@ N.B. This is a breaking change if you use meta on errors.
 * `DocumentInterface::getMeta()` now returns a `Meta` instance instead of a plain array. If no meta is present, it returns `null`. All implementations have been updated to reflect these changes.
 N.B. This is a minor breaking change if you use meta on documents.
 * `DocumentInterface::getJsonapi()` now returns a `Jsonapi` instance instead of a plain array. If no jsonapi is present, it returns `null`. All implementations have been updated to reflect these changes.
+* Parameters for `ItemInterface::setRelation()` have changed to include an optional `Meta` object.
 
 ### Removed